# optimize event loading